Moved remaining ancillary functions from main.go
