+++ main.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"strings"
 )
 
 var (
@@ -26,53 +25,6 @@ var (
 	testUserName     string
 	userCreationPath string
 )
-
-func IgnoreUserLDAPUserObjects(s []LDAPUserObject, e string) bool {
-	for _, a := range s {
-		if a.Uid == e {
-			return true
-		}
-	}
-	return false
-}
-func IgnoreUser(s []string, e string) bool {
-	for _, a := range s {
-		if a == e {
-			return true
-		}
-	}
-	return false
-}
-func TrimSuffix(s, suffix string) string {
-	if strings.HasSuffix(s, suffix) {
-		s = s[:len(s)-len(suffix)]
-	}
-	return s
-}
-func SyncLDAPToConsul(userClass string, usersSet []LDAPUserObject, noop bool, c *ConsulClient, conf Configuration) {
-	namespace := fmt.Sprintf("%s/%s/", conf.Consul.Namespace, userClass)
-	globalAdminsConsul, _, _ := c.client.Keys(namespace, "/", nil)
-	for _, consulAdminUser := range globalAdminsConsul {
-		keyPath := consulAdminUser
-		consulAdminUser = TrimSuffix(consulAdminUser, "/")
-		usernameSplit := strings.Split(consulAdminUser, "/")
-		consulAdminUsername := usernameSplit[len(usernameSplit)-1]
-		ignoreConsulUser := IgnoreUserLDAPUserObjects(usersSet, consulAdminUsername)
-		if consulAdminUsername == userClass {
-			continue
-		}
-		if ignoreConsulUser == false {
-			if noop == false {
-				log.Printf("Removing %s from %s", consulAdminUsername, userClass)
-				log.Printf("KeyPath %s", keyPath)
-				c.client.DeleteTree(keyPath, nil)
-			} else {
-				log.Printf("Should remove %s from %s", consulAdminUsername, userClass)
-			}
-		}
-
-	}
-}
 
 func parseFlags() {
 	flag.StringVar(&testUserName, "testUserName", "", "Test UserName for creating a user. Will be removed, for debugging only.")
