+++ stringutils.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"strings"
+)
+
+func TrimSuffix(s, suffix string) string {
+	if strings.HasSuffix(s, suffix) {
+		s = s[:len(s)-len(suffix)]
+	}
+	return s
+}
